src/python/yamato/requirements.txt
#Adafruit-DHT==1.4.0
appdirs==1.4.4
asn1crypto==1.4.0
attrs==19.3.0
black==19.10b0
CacheControl==0.12.6
#cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
#cryptography==3.2.1
distlib==0.3.1
distro==1.5.0
html5lib==1.1
idna==2.10
msgpack==1.0.0
ordered-set==4.0.2
packaging==20.4
paho-mqtt==1.5.1
pathspec==0.8.0
pep517==0.9.1
ply==3.11
progress==1.5
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytoml==0.1.21
pytz==2021.1
#regex==2020.5.14
requests==2.24.0
resolvelib==0.5.1
retrying==1.3.3
RPi.GPIO==0.7.1
six==1.15.0
systemd~=0.17.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.10
webencodings==0.5.1
